+++ temboardui/__main__.py
@@ -1,7 +1,9 @@
 # coding: utf-8
 
+import imp
 import logging.config
 import os
+import pkg_resources
 import socket
 import sys
 from argparse import (
@@ -20,6 +22,7 @@ from .toolkit import taskmanager
 from .toolkit import validators as v
 from .toolkit.app import define_core_arguments
 from .toolkit.configuration import OptionSpec
+from .toolkit.errors import UserError
 from .toolkit.proctitle import ProcTitleManager
 from .toolkit.services import (
     Service,
@@ -90,7 +93,7 @@ def bootstrap_tornado_app(app, config):
     return app
 
 
-def finalize_tornado_app(app, config):
+def finalize_tornado_app(app, config, plugins):
     base_path = os.path.dirname(__file__)
     handlers = [
         (r"/css/(.*)", tornado.web.StaticFileHandler, {
@@ -107,7 +110,7 @@ def finalize_tornado_app(app, config):
         })
     ]
 
-    config.plugins = legacy_enable_plugins(app, config.temboard['plugins'])
+    config.plugins = legacy_enable_plugins(app, plugins)
     # Append rules *after* plugins because plugins shares same namespace for
     # static rules, i.e. /js/.* is a fallback for /js/dashboard/.*.
     app.add_rules(handlers)
@@ -255,6 +258,31 @@ class TemboardApplication(BaseApplication):
     REPORT_URL = "https://github.com/dalibo/temboard/issues"
     VERSION = __version__
 
+    def filter_plugins(self):
+        # Returns two lists of plugin names from config.temboard.plugins, the
+        # first list is legacy plugins, the second is external plugins.
+
+        legacy_plugins = []
+        ep_plugins = []
+
+        path = os.path.dirname(__file__) + '/plugins'
+
+        logger.debug("Looping over all plugins.")
+        for name in self.config.temboard.plugins:
+            if any(pkg_resources.iter_entry_points(self.with_plugins, name)):
+                logger.debug("%s is a new plugin.", name)
+                ep_plugins.append(name)
+            else:
+                try:
+                    fp, _, _ = imp.find_module(name, [path])
+                except ImportError:
+                    raise UserError("Missing plugin: %s." % name)
+                else:
+                    legacy_plugins.append(name)
+                    logger.debug("%s is a legacy plugin.", name)
+
+        return legacy_plugins, ep_plugins
+
     def apply_config(self):
         bootstrap_tornado = not hasattr(self, 'webapp')
         if bootstrap_tornado:
@@ -264,9 +292,13 @@ class TemboardApplication(BaseApplication):
             self.webapp.executor = ThreadPoolExecutor(12)
             self.webapp.temboard_app = self
 
+        legacy_plugins, ep_plugins = self.filter_plugins()
+        self.config.temboard.plugins = ep_plugins
+
         super(TemboardApplication, self).apply_config()
 
-        finalize_tornado_app(app, self.config)
+        finalize_tornado_app(app, self.config, plugins=legacy_plugins)
+
         self.webapp.engine = configure_db_session(self.config.repository)
 
     def main(self, argv, environ):
@@ -345,7 +377,10 @@ class TemboardApplication(BaseApplication):
             webservice.run()
 
 
-main = TemboardApplication(specs=list_options_specs(), with_plugins=None)
+main = TemboardApplication(
+    specs=list_options_specs(),
+    with_plugins="temboard.plugins",
+)
 
 
 if __name__ == "__main__":